% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{proc_benchmarking}
\alias{proc_tsraking}
\alias{macro_gseriestsbalancing}
\title{Alias de fonction}
\usage{
proc_benchmarking(...)

proc_tsraking(...)

macro_gseriestsbalancing(...)
}
\arguments{
\item{...}{Arguments de la fonction correspondante.}
}
\description{
\code{\link[=proc_benchmarking]{proc_benchmarking()}} est un alias de \code{\link[=benchmarking]{benchmarking()}}

\code{\link[=proc_tsraking]{proc_tsraking()}} est un alias de \code{\link[=tsraking]{tsraking()}}

\code{\link[=macro_gseriestsbalancing]{macro_gseriestsbalancing()}} est un alias de \code{\link[=tsbalancing]{tsbalancing()}}
}
\details{
Le nom de ces alias rappelle l'origine SAS\eqn{^\circledR}{®} de la fonction correspondante
(PROC BENCHMARKING, PROC TSRAKING et macro \emph{\strong{GSeriesTSBalancing}} dans G-Séries 2.0). Ces alias
assurent également la compatibilité avec les premières versions de développement de la librairie R.

Voir la fonction correspondante pour des exemples ou la description des arguments et de la valeur
de retour :
\itemize{
\item \code{\link[=benchmarking]{benchmarking()}} pour l'alias \code{\link[=proc_benchmarking]{proc_benchmarking()}}
\item \code{\link[=tsraking]{tsraking()}} pour l'alias \code{\link[=proc_tsraking]{proc_tsraking()}}
\item \code{\link[=tsbalancing]{tsbalancing()}} pour l'alias \code{\link[=macro_gseriestsbalancing]{macro_gseriestsbalancing()}}
}
}
\keyword{internal}
