% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Coerce a \code{\link[=fr-package]{fr_tdr}} object into a data frame}
\usage{
as_data_frame(x, ...)
}
\arguments{
\item{x}{a \code{\link[=fr-package]{fr_tdr}} object}

\item{...}{ignored}
}
\value{
a data frame
}
\description{
Coerce a \code{\link[=fr-package]{fr_tdr}} object into a data frame
}
\details{
equivalent to \code{as.data.frame()}
}
