% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{as_fr_field}
\alias{as_fr_field}
\title{Coerce \code{character}, \code{factor}, \code{numeric}, \code{logical}, and \code{Date}
vectors into \code{\link[=fr-package]{fr_field}} objects}
\usage{
as_fr_field(x, ...)
}
\arguments{
\item{x}{a character, factor, numeric, integer, logical, or Date vector}

\item{...}{\code{name}, \code{title}, or \code{description} property (\href{https://specs.frictionlessdata.io/table-schema/#name}{\code{name} is required})}
}
\value{
a \link[=fr-package]{fr_field} object
}
\description{
The supported classes of \code{R} objects are converted to the corresponding frictionless \code{type}:\tabular{ll}{
   \strong{\code{R} class} \tab \strong{\code{fr} type} \cr
   \code{character()}, \code{factor()} \tab \code{string}* \cr
   \code{numeric()}, \code{integer()} \tab \code{number} \cr
   \code{logical()} \tab \code{boolean} \cr
   \code{Date} \tab \code{date} \cr
}
}
\details{
*If a \code{factor()}, the \code{enum} \code{constraint} is set to the levels of the factor in R.
}
\examples{
as_fr_field(1:10, "example_integer") # -> frictionless numeric
as_fr_field((1:10) * 0.1, "example_double") # -> frictionless numeric
as_fr_field(letters, "example_character") # -> frictionless string
as_fr_field(factor(letters), "example_factor") # -> frictionless string with enum constraints
as_fr_field(c(TRUE, FALSE, TRUE), "example_logical") # -> frictionless boolean
as_fr_field(as.Date(c("2023-04-23", "2004-12-31")), "example_date") # -> frictionless date

# an example `fr_field` object
uid <-
  replicate(34, paste0(sample(c(letters, 0:9), 8, TRUE), collapse = "")) |>
  as_fr_field(
    name = "uid",
    title = "Unique Identifier",
  )

# set properties of the `fr_field`
S7::prop(uid, "description")
uid@description <- "Consists of 8 random characters from (a-z) and (0-9)"

# inspect properties of the `fr_field`
uid
str(uid)
}
