% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_fr_tdr}
\alias{as_fr_tdr}
\title{Coerce a data frame into a \code{\link[=fr-package]{fr_tdr}} object}
\usage{
as_fr_tdr(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{required (\code{name}) and optional \href{https://specs.frictionlessdata.io/data-resource/#descriptor}{tabular-data-resource properties} (e.g., \code{path}, \code{version}, \code{title}, \code{homepage}, \code{description})}
}
\value{
a \link[=fr-package]{fr_tdr} object
}
\description{
Coerce a data frame into a \code{\link[=fr-package]{fr_tdr}} object
}
