% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_list}
\alias{as_list}
\title{Coerce a \code{\link[=fr-package]{fr_tdr}} object into a list}
\usage{
as_list(x, ...)
}
\arguments{
\item{x}{a \code{\link[=fr-package]{fr_tdr}} object}

\item{...}{ignored}
}
\value{
a list representing the frictionless metadata descriptor
}
\description{
equivalent to \code{as.list()}
}
\examples{
as_fr_tdr(mtcars, name = "mtcars") |>
  as_list()
}
