% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_tbl_df}
\alias{as_tbl_df}
\title{Coerce a \code{\link[=fr-package]{fr_tdr}} object into a data frame}
\usage{
as_tbl_df(x, ...)
}
\arguments{
\item{x}{a \code{\link[=fr-package]{fr_tdr}}}

\item{...}{ignored}
}
\value{
a tibble
}
\description{
Coerce a \code{\link[=fr-package]{fr_tdr}} object into a data frame
}
