% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{as_vector}
\alias{as_vector}
\title{convert a frictionless \href{https://specs.frictionlessdata.io/table-schema/#field-descriptors}{field} object back to a vector}
\usage{
as_vector(x, ...)
}
\arguments{
\item{x}{a \link[=fr-package]{fr_field} object}

\item{...}{ignored}
}
\value{
the vector stored in \code{x@value}, which is a \code{character}, \code{factor}, \code{numeric}, \code{logical}, or \code{Date} vector
}
\description{
convert a frictionless \href{https://specs.frictionlessdata.io/table-schema/#field-descriptors}{field} object back to a vector
}
