% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{format.fr_field}
\alias{format.fr_field}
\title{fr_field(letters, name = "letters")
fr_field(factor(letters), name = "letters")}
\usage{
\method{format}{fr_field}(x, ...)
}
\description{
fr_field(letters, name = "letters")
fr_field(factor(letters), name = "letters")
}
