% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr-package.R
\docType{package}
\name{fr-package}
\alias{fr}
\alias{fr-package}
\title{fr: Frictionless Standards in R}
\description{
Classes for Frictionless table and field schema as well as tabular-data-resource are provided. Tools for creating, curating, writing, and reading tabular-data-resources are also available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cole-brokamp/fr}
  \item \url{https://cole-brokamp.github.io/fr/}
  \item Report bugs at \url{https://github.com/cole-brokamp/fr/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole@colebrokamp.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID})

}
\keyword{internal}
