% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{fr_boolean}
\alias{fr_boolean}
\title{create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#boolean}{boolean} field}
\usage{
fr_boolean(x, name, ..., parse = FALSE)
}
\arguments{
\item{x}{vector coercable to a logical vector}

\item{name}{character string for name of the field; when read from a tdr_csv, \emph{this} (i.e., not
\code{names(x)}) becomes the column name when in a data.frame or tibble}

\item{...}{other metadata descriptors (\code{title}, \code{description})}

\item{parse}{parse x with \code{as.logical} instead of coercion? requires that x is a character vector}
}
\value{
"fr_field" object
}
\description{
create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#boolean}{boolean} field
}
\examples{
fr_boolean(c(TRUE, FALSE, TRUE), name = "case")
# use parse = TRUE to use `as.logical()` to parse character strings for booleans
fr_boolean(c("true", "False", "TRUE", " "), name = "case", parse = TRUE)
# without parse = TRUE, abbreviations will cause an error
}
