% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{fr_date}
\alias{fr_date}
\title{create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#date}{date} field}
\usage{
fr_date(x, name, ..., format = "\%Y-\%m-\%d", parse = FALSE)
}
\arguments{
\item{x}{vector coercable to a Date vector based on \code{format} (ideally as \code{YYYY-MM-DD})}

\item{name}{character string for name of the field; when read from a tdr_csv, \emph{this} (i.e., not
\code{names(x)}) becomes the column name when in a data.frame or tibble}

\item{...}{other metadata descriptors (\code{title}, \code{description})}

\item{format}{as used in \code{strptime()}, defaults to \href{https://en.wikipedia.org/wiki/ISO_8601}{ISO8601}}

\item{parse}{parse x with \code{datefixR::fix_date_char} instead of coercion? requires that x is a character vector}
}
\value{
"fr_field" object
}
\description{
create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#date}{date} field
}
\examples{
# normal usage
fr_date(c("2023-01-01", "2023-03-01", "2022-11-24"), name = "event_date")
# also takes Dates if already converted
fr_date(as.Date(c("2023-01-01", "2023-03-01", "2022-11-24")), name = "event_date")
# can specify format for non-ISO8601 dates
fr_date(c("2023/01/01", "2023/03/01", "2022/11/24"), name = "event_date", format = "\%Y/\%m/\%d")
# can parse dates in other formats with optionally installed {datefixR} package
fr_date(c("02/05/23", "02/05/21", "02/05/18"), name = "event_date", parse = TRUE)
# parsing is useful for year-month combinations too,
# as day is always imputed as the first of the month
fr_date(c("jan 2020", "feb 2020", "mar 2020"), name = "event_date", parse = TRUE)
fr_date(c("2020-01", "2020-02", "2020-03"), name = "event_date", parse = TRUE)
# parsing with a missing day and month (i.e., just year) will cause an error;
# consider using fr_numeric in this case instead
}
