% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_desc.R
\name{fr_desc}
\alias{fr_desc}
\title{extract the frictionless descriptors for a \code{fr_field} or \code{fr_tdr} object
If the \code{fr_field} object has a \code{constraints} list of length zero, then it
is not included in the returned list}
\usage{
fr_desc(x, ...)
}
\arguments{
\item{x}{a \link[=fr-package]{fr_tdr} or \link[=fr-package]{fr_field} object}

\item{...}{ignored}
}
\value{
list of frictionless descriptors
}
\description{
extract the frictionless descriptors for a \code{fr_field} or \code{fr_tdr} object
If the \code{fr_field} object has a \code{constraints} list of length zero, then it
is not included in the returned list
}
