% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{fr_field_boolean}
\alias{fr_field_boolean}
\title{frictionless field: \href{https://specs.frictionlessdata.io/table-schema/#boolean}{boolean} type}
\usage{
fr_field_boolean(x, name, ..., parse = FALSE)
}
\arguments{
\item{x}{vector coercable to a logical vector}

\item{name}{required name metadata descriptor as a string}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> other optional metadata descriptors (\code{title}, \code{description}, \code{example})}

\item{parse}{parse x with \code{as.logical} instead of coercion? requires that x is a character vector}
}
\value{
a \link[=fr-package]{fr_field} vector
}
\description{
frictionless field: \href{https://specs.frictionlessdata.io/table-schema/#boolean}{boolean} type
}
\examples{
fr_field_boolean(c(TRUE, FALSE, TRUE), name = "case")

# use parse = TRUE to use `as.logical()` to parse character strings for booleans
fr_field_boolean(c("true", "False", "TRUE", " "), name = "case", parse = TRUE)

# without parse = TRUE, abbreviations will cause an error
}
