% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{fr_field_string}
\alias{fr_field_string}
\title{frictionless field: \href{https://specs.frictionlessdata.io/table-schema/#string}{string} type}
\usage{
fr_field_string(x, name, ...)
}
\arguments{
\item{x}{vector coerceable to a character vector}

\item{name}{required name metadata descriptor as a string}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> other optional metadata descriptors (\code{title}, \code{description}, \code{example})}
}
\value{
a \link[=fr-package]{fr_field} vector
}
\description{
frictionless field: \href{https://specs.frictionlessdata.io/table-schema/#string}{string} type
}
\examples{
fr_field_string(letters, name = "letters")
fr_field_string(letters, name = "letters", title = "Letters")
# factors are a special type of frictionless string that have an `enum` `constraint`
fr_field_string(letters, name = "letters", constraints = list(enum = letters))
attributes()
}
