% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{fr_string}
\alias{fr_string}
\title{create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#string}{string} field}
\usage{
fr_string(x, name, ...)
}
\arguments{
\item{x}{vector coerceable to a character vector; if x is a factor, its levels will be stored
in ??? TODO factor things...}

\item{name}{character string for name of the field; when read from a tdr_csv, \emph{this} (i.e., not
\code{names(x)}) becomes the column name when in a data.frame or tibble}

\item{...}{other metadata descriptors (see \code{fr_field()} for all possibilities)}
}
\value{
a \link[=fr-package]{fr_field} vector
}
\description{
create a frictionless \href{https://specs.frictionlessdata.io/table-schema/#string}{string} field
}
\examples{
fr_string(letters, name = "letters")
fr_string(letters, name = "letters", title = "Letters")
}
