% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseries-package.R
\docType{package}
\name{gseries-package}
\alias{gseries}
\alias{gseries-package}
\title{gseries: Improve the Coherence of Your Time Series Data}
\description{
'R' version of 'G-Series', Statistics Canada's generalized system devoted to the benchmarking and reconciliation of time series data. The methods used in 'G-Series' essentially come from Dagum, E. B., and P. Cholette (2006) \doi{10.1007/0-387-35439-5}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://StatCan.github.io/gensol-gseries/en/}
  \item \url{https://StatCan.github.io/gensol-gseries/fr/}
  \item Report bugs at \url{https://github.com/StatCan/gensol-gseries/issues/}
}

}
\author{
\strong{Maintainer}: Michel Ferland \email{michel.ferland@statcan.gc.ca}

Authors:
\itemize{
  \item Michel Ferland \email{michel.ferland@statcan.gc.ca}
}

Other contributors:
\itemize{
  \item Statistics Canada [copyright holder, funder]
}

}
\keyword{internal}
