% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{is_fr_field}
\alias{is_fr_field}
\title{Test if an object is a \code{\link[=fr-package]{fr_field}} object}
\usage{
is_fr_field(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
\code{TRUE} if object is a \link[=fr-package]{fr_field} object, \code{FALSE} otherwise
}
\description{
Test if an object is a \code{\link[=fr-package]{fr_field}} object
}
\examples{
is_fr_field(letters)
is_fr_field(as_fr_field(letters, "letters"))
}
