% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graphTable.R
\name{plot_graphTable}
\alias{plot_graphTable}
\title{Générer des graphiques d'étalonnage dans un fichier PDF}
\usage{
plot_graphTable(
  graphTable,
  pdf_file,
  ori_plot_flag = TRUE,
  adj_plot_flag = TRUE,
  GR_plot_flag = TRUE,
  GR_table_flag = FALSE,
  add_bookmarks = TRUE
)
}
\arguments{
\item{graphTable}{(obligatoire)

\emph{Data frame} (object de classe « data.frame ») correspondant au \emph{data frame} de sortie \code{graphTable} de la fonction d'étalonnage.}

\item{pdf_file}{(obligatoire)

Nom (et chemin) du fichier PDF qui contiendra les graphiques d'étalonnage. Le nom doit inclure l'extension de fichier
« .pdf ». Le fichier PDF sera créé dans le répertoire de travail de la session R (tel que renvoyé par \code{getwd()}) si aucun
chemin n'est spécifié. La sécification de \code{NULL} annulerait la création d'un fichier PDF.}

\item{ori_plot_flag, adj_plot_flag, GR_plot_flag, GR_table_flag}{(optionnels)

Arguments logiques (\emph{logical}) indiquant si le type de graphique d'étalonnage correspondant doit être généré ou non. Les trois
premiers types de graphiques sont générés par défaut alors que le quatrième (le tableau des taux de croissance) ne l'est pas.

\strong{Les valeurs par défaut} sont \code{ori_plot_flag = TRUE}, \code{adj_plot_flag = TRUE}, \code{GR_plot_flag = TRUE} et
\code{GR_table_flag = FALSE}.}

\item{add_bookmarks}{Argument logique (\emph{logical}) indiquant si des signets doivent être ajoutés au fichier PDF. Voir \strong{Signets} dans la section
\strong{Détails} pour plus d'informations.

\strong{La valeur par défaut} est \code{add_bookmarks = TRUE}.}
}
\value{
En plus de créer un fichier PDF contenant les graphiques d'étalonnage (sauf si \code{pdf_file = NULL}), cette fonction renvoie également
de manière invisible une liste comprenant les éléments suivants :
\itemize{
\item \code{pdf_name} : Chaîne de caractères (vecteur de type caractère de longueur un) qui contient le nom complet et le chemin du fichier
PDF s'il a été créé avec succès et \code{invisible(NA_character_)} dans le cas contraire ou si \code{pdf_file = NULL} a été spécifié.
\item \code{graph_list} : Liste des graphiques d'étalonnage générés (une par série) comprenant les éléments suivants :
\itemize{
\item \code{name} : Chaîne de caractères décrivant la série (concorde avec le nom du signet dans le fichier PDF).
\item \code{page} : Entier représentant le numéro de séquence du premier graphique de la série dans la séquence complète des graphiques
pour toutes les séries (concorde avec le numéro de page dans le fichier PDF).
\item \code{ggplot_list} : Liste d'objets ggplot (une par graphique ou par page dans le fichier PDF) correspondant aux graphiques
d'étalonnage générés pour la série. Voir la section \strong{Valeur} dans \link{bench_graphs} pour plus de détails.
}
}

Notez que les objets ggplot renvoyés par la fonction peuvent être affichés \emph{manuellement} avec \code{\link[=print]{print()}}, auquel cas certaines mises
à jour des paramètres par défaut du thème ggplot2 sont recommandées afin de produire des graphiques ayant une apparence similaire à
ceux générés dans le fichier PDF (voir la section \strong{Valeur} dans \link{bench_graphs} pour les détails). Gardez également à l'esprit que
ces graphiques sont optimisés pour un format de papier Lettre US en orientation paysage, c.-à-d., 11po de large (27.9cm, 1056px avec
96 PPP) et 8.5po de haut (21.6cm, 816px avec 96 PPP).
}
\description{
Créer un fichier PDF (format de papier lettre US en orientation paysage) contenant des graphiques d'étalonnage pour l'ensemble
des séries contenues dans le \emph{data frame} de sortie \code{graphTable} de la fonction d'étalonnage (\code{\link[=benchmarking]{benchmarking()}} ou
\code{\link[=stock_benchmarking]{stock_benchmarking()}}) spécifié. Quatre types de graphiques d'étalonnage peuvent être générés pour chaque série :
\itemize{
\item \strong{Échelle originale} (argument \code{ori_plot_flag}) - graphique superposé des composantes :
\itemize{
\item Série indicatrice
\item Moyennes de la série indicatrice
\item Série indicatrice corrigée pour le biais (lorsque \eqn{\rho < 1})
\item Série étalonnée
\item Moyennes des étalons
}
\item \strong{Échelle d'ajustement} (argument \code{adj_plot_flag}) - graphique superposé des composantes :
\itemize{
\item Ajustements d'étalonnage
\item Moyennes des ajustements d'étalonnage
\item Ligne du biais (lorsque \eqn{\rho < 1})
}
\item \strong{Taux de croissance} (argument \code{GR_plot_flag}) - diagramme à barres des taux de croissance des séries indicatrice et
étalonnée.
\item \strong{Tableau des taux de croissance} (argument \code{GR_table_flag}) - tableau des taux de croissance des séries indicatrice et
étalonnée.
}

Ces graphiques peuvent être utiles pour évaluer la qualité des résultats de l'étalonnage. N'importe lequel des quatre types de
graphiques d'étalonnage peut être activé ou désactivé à l'aide du drapeau (\emph{flag}) correspondant. Les trois premiers types graphiques
sont générés par défaut alors que le quatrième (le tableau des taux de croissance) ne l'est pas.
}
\details{
Liste des variables du \emph{data frame} \code{graphTable} correspondant à chaque élément des quatre types de graphiques
d'étalonnage:
\itemize{
\item Échelle originale (argument \code{ori_plot_flag})
\itemize{
\item \code{subAnnual} pour la ligne \emph{Indicator Series}
\item \code{avgSubAnnual} pour les segments \emph{Avg. Indicator Series}
\item \code{subAnnualCorrected} pour la ligne \emph{Bias Corr. Indicator Series} (lorsque \eqn{\rho < 1})
\item \code{benchmarked} pour la ligne \emph{Benchmarked Series}
\item \code{avgBenchmark} pour les segments \emph{Average Benchmark}
}
\item Échelle d'ajustement (argument \code{adj_plot_flag})
\itemize{
\item \code{benchmarkedSubAnnualRatio} pour la ligne \emph{BI Ratios (Benchmarked Series / Indicator Series)} \eqn{^{(*)}}{(*)}
\item \code{avgBenchmarkSubAnnualRatio} pour les segments \emph{Average BI Ratios} \eqn{^{(*)}}{(*)}
\item \code{bias} pour la ligne \emph{Bias} (lorsque \eqn{\rho < 1})
}
\item Taux de croissance (argument \code{GR_plot_flag})
\itemize{
\item \code{growthRateSubAnnual} pour les barres \emph{Growth R. in Indicator Series} \eqn{^{(*)}}{(*)}
\item \code{growthRateBenchmarked} pour les barres \emph{Growth R. in Benchmarked Series} \eqn{^{(*)}}{(*)}
}
\item Tableau des taux de croissance (argument \code{GR_table_flag})
\itemize{
\item \code{year} pour la colonne \emph{Year}
\item \code{period} pour la colonne \emph{Period}
\item \code{subAnnual} pour la colonne \emph{Indicator Series}
\item \code{benchmarked} pour la colonne \emph{Benchmarked Series}
\item \code{growthRateSubAnnual} pour la colonne \emph{Growth Rate in Indicator Series} \eqn{^{(*)}}{(*)}
\item \code{growthRateBenchmarked} pour la colonne \emph{Growth Rate in Benchmarked Series} \eqn{^{(*)}}{(*)}
}
}

\eqn{^{(*)}}{(*)} Les \emph{ratios étalons/indicateurs} (« BI ratios ») et les \emph{taux de croissance} (« growth rates ») correspondent
en réalité à des \emph{différences} lorsque \eqn{\lambda = 0} (étalonnage additif).

La fonction utilise les colonnes supplémentaires du \emph{data frame} \code{graphTable} (colonnes non listées dans la section
\strong{Valeur de retour} de \code{\link[=benchmarking]{benchmarking()}} et \code{\link[=stock_benchmarking]{stock_benchmarking()}}), le cas échéant, pour construire les groupes-BY. Voir la
section \strong{Étalonnage de plusieurs séries} de \code{\link[=benchmarking]{benchmarking()}} pour plus de détails.
\subsection{Performance}{

Les deux types de graphiques de taux de croissance, c'est-à-dire le diagramme à barres (\code{GR_plot_flag}) et le tableau (\code{GR_table_flag}),
nécessitent souvent la génération de plusieurs pages dans le fichier PDF, en particulier pour les longues séries mensuelles avec
plusieurs années de données. Cette création de pages supplémentaires ralentit l'exécution de \code{\link[=plot_graphTable]{plot_graphTable()}}. C'est pourquoi
seul le diagramme à barres est généré par défaut (\code{GR_plot_flag = TRUE} et \code{GR_table_flag = FALSE}). La désactivation des deux
types de graphiques de taux de croissance (\code{GR_plot_flag = FALSE} et \code{GR_table_flag = FALSE}) ou la réduction de la taille du
\emph{data frame} d'entrée \code{graphTable} pour les séries très longues (ex., en ne gardant que les années récentes) pourrait ainsi améliorer
le temps d'exécution. Notez également que l'impact de l'étalonnage sur les taux de croissance peut être déduit du graphique dans
l'échelle d'ajustement (\code{adj_plot_flag}) en examinant l'ampleur du mouvement vertical (vers le bas ou vers le haut) des ajustements
d'étalonnage entre deux périodes adjacentes : plus le mouvement vertical est important, plus l'impact sur le taux de croissance
correspondant est important. Le temps d'exécution de \code{\link[=plot_graphTable]{plot_graphTable()}} pourrait donc être reduit, si nécessaire, en ne générant
que les deux premiers types de graphiques et en se concentrant sur le graphique des d'ajustements d'étalonnage pour évaluer la
préservation du mouvement d'une période à l'autre, c'est-à-dire l'impact de l'étalonnage sur les taux de croissance initiaux.
}

\subsection{Thèmes de ggplot2}{

Les graphiques sont générés avec la librairie ggplot2 qui est livrée avec un ensemble pratique de
\href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{thèmes complets} pour l'aspect général des graphiques (avec \code{theme_grey()}
comme thème par défaut). Utilisez la fonction \code{theme_set()} pour changer le thème appliqué aux graphiques générés par \code{\link[=plot_graphTable]{plot_graphTable()}}
(voir les \strong{Exemples}).
}

\subsection{Signets}{

Des signets sont ajoutés au fichier PDF avec \code{xmpdf::set_bookmarks()} lorsque l'argument \ifelse{latex}{\code{add _bookmarks = TRUE}}{
\code{add_bookmarks = TRUE}} (par défault), ce qui nécessite un outil tiers tel que \href{https://www.ghostscript.com/}{Ghostscript} ou
\href{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}{PDFtk}. Voir la section \strong{Installation} dans
\code{vignette("xmpdf", package = "xmpdf")} pour plus de détails.

\strong{Important} : les signets seront ajoutés avec succès au fichier PDF \strong{si et seulement si} \ifelse{latex}{\code{xmpdf::supports 
_set_bookmarks()}}{\code{xmpdf::supports_set_bookmarks()}} renvoie \code{TRUE} \strong{et} l'exécution de \code{xmpdf::set_bookmarks()} est réussie.
Si Ghostscript est installé sur votre machine mais que \code{xmpdf::supports_set_bookmarks()} renvoie toujours \code{FALSE}, essayez de
spécifier le chemin de l'exécutable Ghostscript dans la variable d'environnement \code{R_GSCMD} (ex., \ifelse{latex}{
\code{Sys.setenv(R_GSCMD = "C:/Program Files/.../bin/gswin64c .exe")}}{
\code{Sys.setenv(R_GSCMD = "C:/Program Files/.../bin/gswin64c.exe")}} avec Windows). D'un autre côté, si
\verb{xmpdf::supports_set_bookmarks()\}} renvoie \code{TRUE} mais que vous rencontrez des problèmes (insolubles) avec \code{xmpdf::set_bookmarks()}
(ex., erreur liée à l'exécutable Ghostscript), la création de signets peut être désactivée en spécifiant \code{add_bookmarks = FALSE}.
}
}
\examples{
# Définir le répertoire de travail (pour les fichiers graphiques PDF)
rep_ini <- setwd(tempdir())


# Ventes trimestrielles de voitures et camionnettes (séries indicatrices)
ind_tri <- ts_to_tsDF(
  ts(
    matrix(
      c(
        # Voitures
        1851, 2436, 3115, 2205, 1987, 2635, 3435, 2361, 2183, 2822,
        3664, 2550, 2342, 3001, 3779, 2538, 2363, 3090, 3807, 2631,
        2601, 3063, 3961, 2774, 2476, 3083, 3864, 2773, 2489, 3082,
        # Camionnettes
        1900, 2200, 3000, 2000, 1900, 2500, 3800, 2500, 2100, 3100,
        3650, 2950, 3300, 4000, 3290, 2600, 2010, 3600, 3500, 2100,
        2050, 3500, 4290, 2800, 2770, 3080, 3100, 2800, 3100, 2860),
      ncol = 2
    ),
   start = c(2011, 1),
   frequency = 4,
   names = c("voitures", "camionnettes")
  )
)

# Ventes annuelles de voitures et camionnettes (étalons)
eta_tri <- ts_to_bmkDF(
  ts(matrix(c(# Voitures
              10324, 10200, 10582, 11097, 11582, 11092,
              # Camionnettes
              12000, 10400, 11550, 11400, 14500, 16000),
            ncol = 2),
     start = 2011,
     frequency = 1,
     names = c("voitures", "camionnettes")), 
  ind_frequency = 4)
  

# Étalonnage proportionnel sans correction pour le biais
res_eta <- benchmarking(ind_tri, eta_tri,
                        rho = 0.729, lambda = 1, biasOption = 1,
                        allCols = TRUE,
                        quiet = TRUE)


# Ensemble de graphiques par défaut (les 3 premiers types de graphiques)
\donttest{plot_graphTable(res_eta$graphTable, "graphes_etalonnage.pdf")}

# Utiliser temporairement `theme_bw()` de ggplot2 pour les graphiques
library(ggplot2)
theme_ini <- theme_get()
theme_set(theme_bw())
\donttest{plot_graphTable(res_eta$graphTable, "graphes_etalonnage_bw.pdf")}
theme_set(theme_ini)

# Generer les 4 types de graphiques (incluant le tableau des taux de croissance)
\donttest{plot_graphTable(res_eta$graphTable, "graphes_etalonnage_avec_tableauTC.pdf",
                GR_table_flag = TRUE)}

# Réduire le temps d'exécution en désactivant les deux types de graphiques 
# des taux de croissance
\donttest{plot_graphTable(res_eta$graphTable, "graphes_etalonnage_sans_TC.pdf",
                GR_plot_flag = FALSE)}


# Réinitialiser le répertoire de travail à son emplacement initial
setwd(rep_ini)
}
\seealso{
\link{bench_graphs} \code{\link[=plot_benchAdj]{plot_benchAdj()}} \code{\link[=benchmarking]{benchmarking()}} \code{\link[=stock_benchmarking]{stock_benchmarking()}}
}
