% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fr_tdr.R
\name{read_fr_tdr}
\alias{read_fr_tdr}
\title{read a tabular-data-resource into R}
\usage{
read_fr_tdr(x, ...)
}
\arguments{
\item{x}{path to a \code{tabular-data-resource.yaml} file}

\item{...}{additional options passed onto \code{readr::read_csv()}}
}
\value{
a \link[=fr-package]{fr_tdr} object
}
\description{
read a tabular-data-resource into R
}
