% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-common.R
\name{time_values_conv}
\alias{time_values_conv}
\alias{gs.time2year}
\alias{gs.time2per}
\alias{gs.time2str}
\title{Fonctions de conversion de valeurs de temps}
\usage{
gs.time2year(ts)

gs.time2per(ts)

gs.time2str(ts, sep = "-")
}
\arguments{
\item{ts}{(obligatoire)

Objet de type série chronologique (classe « ts » ou « mts »).}

\item{sep}{(optionnel)

Chaîne de caractères (constante de type caractère) spécifiant le séparateur à
utiliser entre les valeurs d'année et de période.

\strong{La valeur par défaut} est \code{sep = "-"}.}
}
\value{
\code{\link[=gs.time2year]{gs.time2year()}} renvoie un vecteur de nombres entiers correspondant à l'année (l'unité de temps) « la plus proche ».
Cette fonction est l'équivalent de \code{\link[stats:cycle]{stats::cycle()}} pour les valeurs d'unités de temps.

\code{\link[=gs.time2per]{gs.time2per()}} renvoie un vecteur de nombres entiers contenant les valeurs des périodes (les cycles ; voir
\code{\link[stats:cycle]{stats::cycle()}})).

\code{\link[=gs.time2str]{gs.time2str()}} renvoie un vecteur de chaînes de caractères correspondant à \code{gs.time2year(ts)} lorsque
\code{stats::frequency(ts) == 1} ou à \code{gs.time2year(ts)} et \code{gs.time2per(ts)} séparé par \code{sep} dans le cas contraire.
}
\description{
Fonctions de conversion de valeurs de temps utilisées à l'interne par d'autres fonctions de gseries.
}
\examples{
# Série chronologique mensuelle « bidon »
sc_men <- ts(rep(NA, 15), start = c(2019, 1), frequency = 12)
sc_men
gs.time2year(sc_men)
gs.time2per(sc_men)
gs.time2str(sc_men)
gs.time2str(sc_men, sep = "m")

# Série chronologique trimestrielle « bidon »
sc_tri <- ts(rep(NA, 5), start = c(2019, 1), frequency = 4)
sc_tri
gs.time2year(sc_tri)
gs.time2per(sc_tri)
gs.time2str(sc_tri)
gs.time2str(sc_tri, sep = "t")
}
\seealso{
\code{\link[=ts_to_tsDF]{ts_to_tsDF()}} \code{\link[=ts_to_bmkDF]{ts_to_bmkDF()}} \code{\link[=gs.build_proc_grps]{gs.build_proc_grps()}}
}
