% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_bmkDF.R
\name{ts_to_bmkDF}
\alias{ts_to_bmkDF}
\title{Convertir un objet « ts » en \emph{data frame} d'étalons}
\usage{
ts_to_bmkDF(
  in_ts,
  ind_frequency,
  discrete_flag = FALSE,
  alignment = "b",
  bmk_interval_start = 1,
  startYr_cName = "startYear",
  startPer_cName = "startPeriod",
  endYr_cName = "endYear",
  endPer_cName = "endPeriod",
  val_cName = "value"
)
}
\arguments{
\item{in_ts}{(obligatoire)

Objet de type série chronologique (classe « ts » ou « mts ») à convertir.}

\item{ind_frequency}{(obligatoire)

Entier spécifiant la fréquence de la série indicatrice (haute fréquence) à laquelle les étalons (séries de basse fréquence)
sont liés. La fréquence d'une série chronologique correspond au nombre maximum de périodes dans une année (par exemple, 12
pour des données mensuelles, 4 pour des données trimestrielles, 1 pour des données annuelles).}

\item{discrete_flag}{(optionnel)

Argument logique (\emph{logical}) précisant si les étalons correspondent à des valeurs discrètes (points d'ancrage couvrant une
seule période de la série indicatrice, par exemple des stocks de fin d'année) ou non. \code{discrete_flag = FALSE} définit des
étalons non discrets, c'est-à-dire des étalons qui couvrent plusieurs périodes de la série indicatrice (par exemple, des
étalons annuels couvrent 4 trimestres ou 12 mois, des étalons trimestriels couvrent 3 mois, etc.).

\strong{La valeur par défaut} est \code{discrete_flag = FALSE}.}

\item{alignment}{(optionnel)

Caractère identifiant l'alignement des étalons discrets (argument \code{discrete_flag = TRUE}) dans la fenêtre de couverture
de l'intervalle de l'étalon (série de basse fréquence) :
\itemize{
\item \code{alignment = "b"} : début de la fenêtre de l'intervalle de l'étalon (première période)
\item \code{alignment = "e"} : fin de la fenêtre de l'intervalle de l'étalon (dernière période)
\item \code{alignment = "m"} : milieu de la fenêtre de l'intervalle de l'étalon (période du milieu)
}

Cet argument n'a pas d'effet pour les étalons non discrets (\code{discrete_flag = FALSE}).

\strong{La valeur par défaut} est \code{alignment = "b"}.}

\item{bmk_interval_start}{(optionnel)

Entier dans l'intervalle [1 .. \code{ind_frequency}] spécifiant la période (cycle) de la série indicatrice (haute fréquence)
à laquelle commence la fenêtre de l'intervalle de l'étalon (série de basse fréquence). Par exemple, des étalons annuels
correspondant à des années financières définies d'avril à mars de l'année suivante seraient spécifiés avec
\code{bmk_interval_start = 4} pour une série indicatrice mensuelle (\code{ind_frequency = 12}) et \code{bmk_interval_start = 2} pour
une série indicatrice trimestrielle (\code{ind_frequency = 4}).

\strong{La valeur par défaut} est \code{bmk_interval_start = 1}.}

\item{startYr_cName, startPer_cName, endYr_cName, endPer_cName}{(optionnel)

Chaînes de caractères spécifiant le nom des variables (colonnes) numériques dans le \emph{data frame} de sortie qui
définiront la couverture des étalons, c'est-à-dire les identificateurs de l'année et de la période de début et de
fin des étalons.

\strong{Les valeurs par défaut} sont \code{startYr_cName = "startYear"}, \code{startPer_cName = "startPeriod"}
\code{endYr_cName = "endYear"} et \code{endPer_Name = "endPeriod"}.}

\item{val_cName}{(optionnel)

Chaîne de caractères spécifiant le nom de la variable (colonne) dans le \emph{data frame} de sortie qui contiendra les
valeurs des étalons. Cet argument n'a aucun effet pour les objets « mts » (les noms des variables d'étalons sont
automatiquement hérités de l'objet « mts »).

\strong{La valeur par défaut} est \code{val_cName = "value"}.}
}
\value{
La fonction renvoie un \emph{data frame} avec cinq variables ou plus :
\itemize{
\item Année de début de la couverture de l'étalon, type numérique (voir argument \code{startYr_cName})
\item Période de début de la couverture de l'étalon, type numérique (voir argument \code{startPer_cName})
\item Année de fin de la couverture de l'étalon, type numérique (voir argument \code{endYr_cName})
\item Période de fin de la couverture de l'étalon, type numérique (voir argument \code{endPer_cName})
\item Une (objet « ts ») ou plusieurs (objet « mts ») variable(s) de données d'étalons, type numérique
(voir argument \code{val_cName})
}

Note : la fonction renvoie un objet « data.frame » qui peut être explicitement converti en un autre
type d'objet avec la fonction \verb{as*()} appropriée (ex., \code{tibble::as_tibble()} le convertirait en tibble).
}
\description{
Convertir un objet « ts  » (ou « mts ») en un \emph{data frame} d'étalons pour les fonctions d'étalonnage avec cinq variables
(colonnes) ou plus :
\itemize{
\item quatre (4) pour la converture de l'étalon
\item une (1) pour chaque série chronologique d'étalons
}

Pour des étalons discrets (points d'ancrage couvrant une seule période de la série indicatrice, par exemple, des stocks de
fin d'année), spécifiez \code{discrete_flag = TRUE} et \code{alignment = "b"}, \code{"e"} ou \code{"m"}.
}
\examples{
# Séries chronologiques annuelle et trimestrielle
ma_sc_ann <- ts(1:5 * 100, start = 2019, frequency = 1)
ma_sc_ann
ma_sc_tri <- ts(1:5 * 10, start = c(2019, 1), frequency = 4)
ma_sc_tri


# Étalons annuels pour des séries indicatrices mensuelles
ts_to_bmkDF(ma_sc_ann, ind_frequency = 12)

# Étalons annuels pour des série indicatrices trimestrielles
ts_to_bmkDF(ma_sc_ann, ind_frequency = 4)

# Étalons trimestriels pour des séries indicatrices mensuelles
ts_to_bmkDF(ma_sc_tri, ind_frequency = 12)

# Stocks de début d'année pour des séries indicatrices trimestrielles
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 4,
  discrete_flag = TRUE
)

# Stocks de fin de trimestre pour des séries indicatrices mensuelles
ts_to_bmkDF(
  ma_sc_tri, ind_frequency = 12,
  discrete_flag = TRUE, alignment = "e"
)

# Étalons annuels (avril à mars) pour des séries indicatrices ...
# ... mensuelles
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 12,
  bmk_interval_start = 4
)
# ... trimestrielles
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 4,
  bmk_interval_start = 2
)

# Stocks de fin d'année (avril à mars) pour des séries indicatrices ...
# ... mensuelles
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 12,
  discrete_flag = TRUE, alignment = "e", bmk_interval_start = 4
)
# ... trimestrielles
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 4,
  discrete_flag = TRUE, alignment = "e", bmk_interval_start = 2
)

# Nom personnalisé pour la variable (colonne) des étalons
ts_to_bmkDF(
  ma_sc_ann, ind_frequency = 12,
  val_cName = "eta_val"
)

# Séries chronologiques multiples: argument `val_cName` ignoré
# (les noms des colonnes de l'object « mts » sont toujours utilisés)
ts_to_bmkDF(
  ts.union(ser1 = ma_sc_ann, ser2 = ma_sc_ann / 10), ind_frequency = 12,
  val_cName = "nom_de_colonne_inutile"
)
}
\seealso{
\code{\link[=ts_to_tsDF]{ts_to_tsDF()}} \code{\link[=stack_bmkDF]{stack_bmkDF()}} \code{\link[=benchmarking]{benchmarking()}} \code{\link[=stock_benchmarking]{stock_benchmarking()}} \link{time_values_conv}
}
