% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsDF_to_ts.R
\name{tsDF_to_ts}
\alias{tsDF_to_ts}
\title{Fonction réciproque de \code{\link[=ts_to_tsDF]{ts_to_tsDF()}}}
\usage{
tsDF_to_ts(
  ts_df,
  frequency,
  yr_cName = "year",
  per_cName = "period"
)
}
\arguments{
\item{ts_df}{(obligatoire)

\emph{Data frame} (object de classe « data.frame ») à convertir.}

\item{frequency}{(obligatoire)

Entier spécifiant la fréquence de la (des) série(s) à convertir. La fréquence d'une série chronologique correspond
au nombre maximum de périodes dans une année (par exemple, 12 pour des données mensuelles, 4 pour des données trimestrielles,
1 pour des données annuelles).}

\item{yr_cName, per_cName}{(optionnel)

Chaînes de caractères spécifiant le nom des variables (colonnes) numériques dans le \emph{data frame} d'entrée
qui contiennent les identificateurs d'année et de période du point de données.

\strong{Les valeurs par défaut} sont \code{yr_cName = "year"} et \code{per_cName = "period"}.}
}
\value{
La fonction renvoie un objet de type série chronologique (classe « ts » ou « mts »), qui peut être explicitement converti
en un autre type d'objet avec la fonction \verb{as*()} appropriée (ex., \code{tsibble::as_tsibble()} le convertirait en tsibble).
}
\description{
Convertir un \emph{data frame} (non empilé) de séries chronologiques (format de données de \code{\link[=benchmarking]{benchmarking()}} et \code{\link[=stock_benchmarking]{stock_benchmarking()}})
en un objet « ts » (ou « mts »).

Cette fonction est utile pour convertir le \emph{data frame} renvoyé par un appel à \code{\link[=benchmarking]{benchmarking()}} ou \code{\link[=stock_benchmarking]{stock_benchmarking()}}
en un objet « ts », où une ou plusieurs séries ont été étalonnées en mode de traitement \emph{non groupes-BY}. Les
\emph{data frame} empilés de séries chronologiques associées à des exécutions en mode \emph{groupes-BY} doivent d'abord être
\emph{désempilés} avec \code{\link[=unstack_tsDF]{unstack_tsDF()}}.
}
\examples{
# Série chronologique trimestrielle initiale (série indicatrice à étalonner)
sc_tri <- ts(
  c(1.9, 2.4, 3.1, 2.2, 2.0, 2.6, 3.4, 2.4, 2.3),
  start = c(2015, 1), frequency = 4
)

# Série chronologique annuelle (étalons)
sc_ann <- ts(c(10.3, 10.2), start = 2015, frequency = 1)


# Étalonnage proportionnel
res_eta <- benchmarking(
  ts_to_tsDF(sc_tri),
  ts_to_bmkDF(sc_ann, ind_frequency = 4),
  rho = 0.729, lambda = 1, biasOption = 3,
  quiet = TRUE
)

# Séries chronologiques initiale et finale (étalonnée) - objects « ts » 
sc_tri
tsDF_to_ts(res_eta$series, frequency = 4)


# Étalonnage proportionnel de stocks de fin d'année - plusieurs (3) séries 
# traitées avec l'argument `by` (en mode groupes-BY)
sc_tri2 <- ts.union(ser1 = sc_tri,     ser2 = sc_tri * 100, ser3 = sc_tri * 10)
sc_ann2 <- ts.union(ser1 = sc_ann / 4, ser2 = sc_ann * 25,  ser3 = sc_ann * 2.5)
res_eta2 <- stock_benchmarking(
  stack_tsDF(ts_to_tsDF(sc_tri2)),
  stack_bmkDF(ts_to_bmkDF(
    sc_ann2, ind_frequency = 4,
    discrete_flag = TRUE, alignment = "e")
  ),
  rho = 0.729, lambda = 1, biasOption = 3,
  by = "series",
  quiet = TRUE
)

# Séries chronologiques initiales et finales (étalonnées) - objects « mts » 
sc_tri2
tsDF_to_ts(unstack_tsDF(res_eta2$series), frequency = 4)
}
\seealso{
\code{\link[=ts_to_tsDF]{ts_to_tsDF()}} \code{\link[=unstack_tsDF]{unstack_tsDF()}} \code{\link[=benchmarking]{benchmarking()}} \code{\link[=stock_benchmarking]{stock_benchmarking()}}
}
