% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Coerce a \code{\link[=fr-package]{fr_tdr}} object into a data frame}
\usage{
as_data_frame(x, ...)
}
\arguments{
\item{x}{a \code{\link[=fr-package]{fr_tdr}} object}

\item{...}{ignored}
}
\value{
a data frame
}
\description{
Equivalent to \code{as.data.frame()}; directly using \code{tibble::as_tibble()}
also works because its input is first coerced with \code{as.data.frame()}
}
\examples{
as_fr_tdr(mtcars, name = "mtcars") |>
  as_data_frame()
}
