% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_tdr.R
\name{as_fr_tdr}
\alias{as_fr_tdr}
\title{Coerce a data frame into a \code{\link[=fr-package]{fr_tdr}} object}
\usage{
as_fr_tdr(x, ...)
}
\arguments{
\item{x}{a data.frame}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> required (\code{name}) and optional \href{https://specs.frictionlessdata.io/data-resource/#descriptor}{tabular-data-resource properties} (e.g., \code{path}, \code{version}, \code{title}, \code{homepage}, \code{description})}
}
\value{
a \link[=fr-package]{fr_tdr} object
}
\description{
Coerce a data frame into a \code{\link[=fr-package]{fr_tdr}} object
}
\details{
Use the \code{.template} argument to provide a template \code{\link[=fr-package]{fr_tdr}} object from which
table-specific (i.e. "name", "version", "title", "homepage", "description")
and field-specific metadata will be copied; note that all metadata provided
in \code{...} will be ignored if this argument is provided
}
\examples{
as_fr_tdr(mtcars, name = "mtcars")
as_fr_tdr(mtcars, name = "mtcars")@schema
}
