% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_field.R
\name{as_vector}
\alias{as_vector}
\title{Coerce a \code{\link[=fr-package]{fr_field}} object into a vector}
\usage{
as_vector(x, ...)
}
\arguments{
\item{x}{a \code{\link[=fr-package]{fr_field}} object}

\item{...}{ignored}
}
\value{
depending on the \code{type} property of the object, a \code{character}, \code{factor}, \code{numeric}, \code{logical}, or \code{Date} vector
}
\description{
Coerce a \code{\link[=fr-package]{fr_field}} object into a vector
}
