% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fr_schema.R
\name{fr_schema}
\alias{fr_schema}
\title{extract the schema from a \code{fr_tdr} as a list}
\usage{
fr_schema(x, ...)
}
\arguments{
\item{x}{a \link[=fr-package]{fr_tdr} object}

\item{...}{ignored}
}
\value{
a list of fields, where each entry is a list of field-specific metadata descriptors
}
\description{
extract the schema from a \code{fr_tdr} as a list
}
