% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fr_tdr.R
\name{write_fr_tdr}
\alias{write_fr_tdr}
\title{write a fr_tdr object to disk}
\usage{
write_fr_tdr(x, dir)
}
\arguments{
\item{x}{a \link[=fr-package]{fr_tdr} object to write to disk}

\item{dir}{path to directory where tabular-data-resource folder will be created}
}
\value{
x (invisibly)
}
\description{
The \code{name} property of the \code{fr_tdr} object is used to write a frictionless tabular-data-resource to disk.  For example, if \code{name = "my_data"}, then a folder named \code{my_data} would be created with (1) \code{my_data.csv} and (2) \code{tabular-data-resource.yaml}.
}
